전체 재점검에서 찾은 버그 4건 수정
1. 달력 [오늘] 버튼이 한국시간 자정~오전 9시 사이에 '어제'를 골랐습니다.
   toISOString() 이 UTC 라서입니다. 프로토타입의 todayISO()(현지 시간)로
   바꿨습니다. (전체 현황 달력 + 크게 보기 달력 두 곳)

2. 성과지표가 소수점을 받지 못했습니다. '만족도 4.5점' 같은 목표가 실제로
   있고 프로토타입도 Number() 로 허용했는데, 서버가 정수만 받아 형식
   오류가 났습니다. 목표·실적 모두 Float 로 바꿨습니다(마이그레이션 포함).
   화면에서 4.5점 등록 → 실적 0.5 입력 → '0.5 / 4.5점' 누계까지 확인.

3. 공고 화면의 쪽당 개수 저장이 localStorage 를 try 없이 써서,
   사파리 프라이빗 모드에서 화면이 통째로 깨질 수 있었습니다.

4. 서버 형식 오류(pydantic)의 detail 이 배열이라 화면에
   "[object Object]" 로 보였습니다. 사람이 읽을 문구로 바꿨습니다.

검증: pytest 62건(소수점 2건 추가) · 프로토타입 대조 통과

+++ backend/app/api/entries.py
@@ -68,7 +68,7 @@ class SpendIn(BaseModel):
 
 class EntryIn(BaseModel):
     spends: list[SpendIn] = Field(default_factory=list)
-    kpi: dict[str, int] = Field(default_factory=dict)
+    kpi: dict[str, float] = Field(default_factory=dict)
     act: str = ""
     issue: str = ""
     plan: str = ""
